requirements.txt
# --- Núcleo do projeto (ETL + UI) ---
pandas>=2.1
openpyxl>=3.1
streamlit>=1.36
python-dotenv>=1.0
pypdf>=4.2

# --- RAG com LangChain + Qdrant (sem OpenAI) ---
langchain>=0.2.11
langchain-community>=0.2.10
langchain-qdrant>=0.1.0
qdrant-client>=1.15.1
sentence-transformers>=2.7.0

# (opcional) utilidades
tqdm>=4.66
